api/requirements.txt
fastapi==0.98.0
icecream==2.1.3
langchain==0.0.271
openai==0.27.8
python-dotenv==1.0.0
pytz==2023.3.post1
uvicorn==0.21.1